WV-10-08: delete _toyama file and _notifier file for debuggingg
